+++ hydrocron/db/load_data.py
@@ -4,9 +4,6 @@ the appropriate DynamoDB table
 """
 import logging
 import os
-import base64
-import json
-import requests
 
 import boto3
 import earthaccess
@@ -14,6 +11,7 @@ from botocore.exceptions import ClientError
 
 from hydrocron.db import HydrocronTable
 from hydrocron.db.io import swot_reach_node_shp
+from hydrocron.utils import connection
 from hydrocron.utils import constants
 
 
@@ -48,93 +46,35 @@ def lambda_handler(event, _):  # noqa: E501 # pylint: disable=W0613
         start_date,
         end_date)
 
+    lambda_client = boto3.client('lambda')
+
     for granule in new_granules:
-        s3_resource = setup_s3connection()
-        items = read_data(granule, obscure_data, s3_resource)
+        granule_path = granule.data_links(access='direct')[0]
 
-        dynamo_resource = setup_dynamoconnection()
-        load_data(dynamo_resource, table_name, items)
+        event2 = ('{"body": {"granule_path": "'
+                  + granule_path + '","obscure_data": "'
+                  + obscure_data + '","table_name": "'
+                  + table_name + '"}}')
+
+        lambda_client.invoke(
+            FunctionName=os.environ['GRANULE_LAMBDA_FUNCTION_NAME'],
+            InvocationType='Event',
+            Payload=event2)
 
 
-def retrieve_credentials():
-    """Makes the Oauth calls to authenticate with EDS and return a set of s3
-    same-region, read-only credntials.
+def granule_handler(event, _):
     """
-    login_resp = requests.get(
-        constants.S3_CREDS_ENDPOINT,
-        allow_redirects=False,
-        timeout=5
-    )
-    login_resp.raise_for_status()
-
-    auth = f"{os.environ['EARTHDATA_USERNAME']}:{os.environ['EARTHDATA_PASSWORD']}"
-    encoded_auth = base64.b64encode(auth.encode('ascii'))
-
-    auth_redirect = requests.post(
-        login_resp.headers['location'],
-        data={"credentials": encoded_auth},
-        headers={"Origin": constants.S3_CREDS_ENDPOINT},
-        allow_redirects=False,
-        timeout=5
-    )
-    auth_redirect.raise_for_status()
-
-    final = requests.get(
-        auth_redirect.headers['location'],
-        allow_redirects=False,
-        timeout=5
-    )
-
-    results = requests.get(
-        constants.S3_CREDS_ENDPOINT,
-        cookies={'accessToken': final.cookies['accessToken']},
-        timeout=5
-    )
-    results.raise_for_status()
-
-    return json.loads(results.content)
-
-
-def setup_s3connection():
+    Second Lambda entrypoint for loading individual granules
     """
-    Set up S3 resource connections
+    granule_path = event['body']['granule_path']
+    obscure_data = event['body']['obscure_data']
+    table_name = event['body']['table_name']
 
-    Returns
-    -------
-    s3_resource : S3 resource
-    """
+    s3_resource = connection.s3_resource
+    items = read_data(granule_path, obscure_data, s3_resource)
 
-    creds = retrieve_credentials()
-
-    s3_session = boto3.session.Session(
-        aws_access_key_id=creds['accessKeyId'],
-        aws_secret_access_key=creds['secretAccessKey'],
-        aws_session_token=creds['sessionToken'],
-        region_name='us-west-2')
-
-    s3_resource = s3_session.resource('s3')
-
-    return s3_resource
-
-
-def setup_dynamoconnection():
-    """
-    Set up DynamoDB resource connections
-
-    Returns
-    -------
-    dynamo_resource : HydrocronDB
-
-    """
-
-    dyn_session = boto3.session.Session()
-
-    if endpoint_url := os.getenv('HYDROCRON_dynamodb_endpoint_url'):
-        dyndb_resource = dyn_session.resource('dynamodb', endpoint_url=endpoint_url)
-    else:
-        dyndb_resource = dyn_session.resource('dynamodb')
-
-    return dyndb_resource
+    dynamo_resource = connection.dynamodb_resource
+    load_data(dynamo_resource, table_name, items)
 
 
 def find_new_granules(collection_shortname, start_date, end_date):
@@ -162,14 +102,14 @@ def find_new_granules(collection_shortname, start_date, end_date):
     return results
 
 
-def read_data(granule, obscure_data, s3_resource=None):
+def read_data(granule_path, obscure_data, s3_resource=None):
     """
     Read data from shapefiles
 
     Parameters
     ----------
-    granule : Granule
-        the granule to unpack
+    granule_path : string
+        the S3 url to the granule to unpack
     obscure_data : boolean
         whether to obscure the data on load
     s3_resource : boto3 session resource
@@ -178,7 +118,7 @@ def read_data(granule, obscure_data, s3_resource=None):
     -------
     items : the unpacked granule data
     """
-    granule_path = granule.data_links(access='direct')[0]
+    items = {}
 
     if 'Reach' in granule_path:
         items = swot_reach_node_shp.read_shapefile(
@@ -203,8 +143,10 @@ def load_data(dynamo_resource, table_name, items):
 
     Parameters
     ----------
-    hydrocron_table : HydrocronTable
-        The table to load data into
+    dynamo_resource : Resource
+        Dynamo resource
+    table_name : String
+        The name of the table
     items : Dictionary
         The unpacked granule to load
     """
@@ -219,13 +161,11 @@ def load_data(dynamo_resource, table_name, items):
     if hydrocron_table.table_name == constants.SWOT_REACH_TABLE_NAME:
 
         for item_attrs in items:
-            # write to the table
             hydrocron_table.add_data(**item_attrs)
 
     elif hydrocron_table.table_name == constants.SWOT_NODE_TABLE_NAME:
 
         for item_attrs in items:
-            # write to the table
             hydrocron_table.add_data(**item_attrs)
 
     else: