+++ terraform/hydrocron-lambda.tf
@@ -15,6 +15,7 @@ locals {
   ecr_image_tag                    = element(local.ecr_image_name_and_tag, 1)
   timeseries_function_name         = "${local.aws_resource_prefix}-timeseries-lambda"
   load_data_function_name          = "${local.aws_resource_prefix}-load_data-lambda"
+  load_granule_function_name       = "${local.aws_resource_prefix}-load_granule-lambda"
 }
 
 resource aws_ecr_repository "lambda-image-repo" {
@@ -71,10 +72,9 @@ resource "aws_lambda_permission" "allow_hydrocron-timeseries" {
   action        = "lambda:InvokeFunction"
   function_name = aws_lambda_function.hydrocron_lambda_timeseries.function_name
   principal     = "apigateway.amazonaws.com"
-
-  # The "/*/*/*" portion grants access from any method on any resource
-  # within the API Gateway REST API.
-  source_arn = "${aws_api_gateway_rest_api.hydrocron-api-gateway.execution_arn}/*/*/*"
+  # The /* part allows invocation from any stage, method and resource path
+  # within API Gateway.
+  source_arn = "${aws_api_gateway_rest_api.hydrocron-api-gateway.execution_arn}/*"
 }
 
 
@@ -94,6 +94,36 @@ resource "aws_lambda_function" "hydrocron_lambda_load_data" {
     variables = {
       EARTHDATA_USERNAME = data.aws_ssm_parameter.edl_username.value
       EARTHDATA_PASSWORD = data.aws_ssm_parameter.edl_password.value
+      GRANULE_LAMBDA_FUNCTION_NAME = aws_lambda_function.hydrocron_lambda_load_granule.function_name
     }
   }
+}
+
+
+resource "aws_lambda_function" "hydrocron_lambda_load_granule" {
+  package_type = "Image"
+  image_uri    = "${aws_ecr_repository.lambda-image-repo.repository_url}:${data.aws_ecr_image.lambda_image.image_tag}"
+  image_config {
+    command = ["hydrocron.db.load_data.granule_handler"]
+  }
+  function_name = local.load_granule_function_name
+  role          = aws_iam_role.hydrocron-lambda-load-granule-role.arn
+  timeout       = 600
+  memory_size   = 512
+
+  tags = var.default_tags
+  environment {
+    variables = {
+      EARTHDATA_USERNAME = data.aws_ssm_parameter.edl_username.value
+      EARTHDATA_PASSWORD = data.aws_ssm_parameter.edl_password.value
+    }
+  }
+}
+
+resource "aws_lambda_permission" "allow_lambda" {
+  statement_id  = "AllowExecutionFromLambda"
+  action        = "lambda:InvokeFunction"
+  function_name = aws_lambda_function.hydrocron_lambda_load_granule.function_name
+  principal     = "s3.amazonaws.com"
+  source_arn = aws_lambda_function.hydrocron_lambda_load_data.arn
 }