+++ CHANGELOG.md
@@ -6,10 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+    - Issue 88 - There are no CloudWatch logs for the API Gateway
+    - Issue 75 - Update log messaging format
+    - Issue 60 - Encapsulate DyanmoDB under a single shared module
+    - Issue 60 - Improved error handling
 ### Changed
 ### Deprecated 
 ### Removed
+    - Issue 39 - Clean up code, removed comments and unused parameters
 ### Fixed
+    - Issue 44 - Load data lambda only loads the first granule found in the time range
 ### Security
 
 ## [1.0.0]